pyproject.toml
[tool.poetry]
authors = ["Pritam Dey <pritam.pritamdey.984@gmail.com>"]
description = "Interact with LLMs"
name = "interact"
readme = "README.md"
version = "0.5.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry.dependencies]
numpy = "^2.0.1"
openai = "^1.12"
pillow = "^10.0"
python = "^3.10"

# extras
faiss-cpu = {version = "^1.8.0", optional = true}

[tool.poetry.extras]
faiss = ["faiss-cpu"]

[tool.poetry.dev-dependencies]
pytest = "^8.2"
python-dotenv = "*"
sphinx = "^7.0"
sphinx-autobuild = "2021.3.14"
sphinx-rtd-theme = "1.3.0"
sphinxcontrib-mermaid = {git = "https://github.com/mgaitan/sphinxcontrib-mermaid.git", rev = "7922ed7"}

[tool.pytest.ini_options]
addopts = ""
minversion = "8.0"
testpaths = ["tests"]
